pyproject.toml
[project]
name = "libstat17"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "regex (>=2025.11.3,<2026.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "vk-api (>=11.10.0,<12.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "snakemake (>=9.14.5,<10.0.0)",
]

[tool.poetry]
packages = [
    {include = "lib"},
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "mypy (>=1.19.1,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "pandas-stubs (>=2.3.3.251219,<3.0.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
    "types-regex (>=2025.11.3.20251106,<2026.0.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
